requirements.txt
flask
psycopg2-binary
flask_sqlalchemy  # <--- Añadir esta línea
werkzeug          # <--- Añadir esta (para el hash de contraseñas)
requests